Day_0102
반복문 if문 연습

+++ Day_1231/src/Exam_10.java
@@ -24,7 +24,7 @@ public class Exam_10 {
 			}
 		}
 		scn.close();
-		System.out.println("«’∞Ë = "+sum+"\t∆Ú±’= "+(sum/cnt));
+		System.out.println("«’∞Ë = "+sum+"\t∆Ú±’= "+((double)sum/cnt));
 		
 	}
 }
